Add code for navbar
